refactor: SnackBarContainerReducer -> snackBarReducer 이름 변경
